Add acme_sh_staging variable
Rather than making users of this role manage the acme_sh_issue_suffix
variable to enable and disable use of the staging server, implement a
default boolean variable (acme_sh_staging) to add (or not) the
"--staging" flag to the acme.sh command.

Users can still use acme_sh_issue_suffix to selectively set --staging
for some domains.

This changeset updates the README (which was out of date anyway,
referring to the use of acme_sh_env to control staging).

It also bumps the required ansible version to 1.9 (required for the
ternary filter).